exercicios 3 e 4
