+++ semana14/node-e-package.json/Exercicio2/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "compilerOptions": {
+      "noImplicitAny": true,
+      "removeComments": true,
+      "preserveConstEnums": true,
+      "outDir": "build",
+      "sourceMap": true,
+      "target": "es6",
+      "module": "commonjs"
+  },
+  "include": [
+      "src/**/*"
+  ],
+  "exclude": [
+      "node_modules"
+  ]
+}
